single border

+++ lua/plugins.lua
@@ -172,7 +172,7 @@ return require("packer").startup({
             moved_sym = "→", -- The symbol for a plugin which was moved (e.g. from opt to start)
             header_sym = "━", -- The symbol for the header line in packer's display
             show_all_info = true, -- Should packer show all update details automatically?
-            prompt_border = "double", -- Border style of prompt popups.
+            prompt_border = "single", -- Border style of prompt popups.
         },
         profile              = {
             enable = false,
